refactor: organize reading components into dedicated subdirectory
- Move ReadingCard component to src/app/components/readings/
- Move placeholderUtils.ts to src/app/components/readings/
- Create proper test directory structure under readings/
- Update import paths in affected files
- Mark 'Organize components' task as completed in TODO.md
