+++ src/app/components/readings/placeholderUtils.ts
